Extract Point2D / SortPoints2D / ConvexHullCCW to shared header
Three copies of the same Andrew's monotone-chain hull algorithm lived
in the tree (Frustum.cpp file-scope, Frustum::SetCustom2DHull's local
lambda, and ZzzLodTerrain.cpp file-scope), all with identical logic
and slightly different identifiers. Factored them into
Camera/ConvexHull2D.h as inline free functions with a shared Point2D.

Per Coding Rule 4 (Write Reusable Code).
